first

+++ app/src/main/java/edu/ucsd/cse110/lab6/Exercises.java
@@ -16,19 +16,19 @@ public class Exercises {
         for (int number : numbers) {
             sum += number;
         }
-        return sum / numbers.size();
+        return ((double)sum) / ((double)numbers.size());
     }
 
     public static double standardDeviation(List<Integer> numbers) {
         double mean = mean(numbers);
 
-        int deviationSum = 0;
+        double deviationSum = 0;
         for (int number : numbers) {
             deviationSum += Math.pow(mean - number, 2);
         }
 
-        int length = numbers.size();
-        float variance = deviationSum / length;
+        double length = numbers.size();
+        double variance = deviationSum / length;
         return Math.sqrt(variance);
     }
